+++ modules/seqexec/model/src/main/scala/seqexec/model/operations.scala
@@ -3,7 +3,6 @@
 
 package seqexec.model
 
-import seqexec.model.Model.Step
 import seqexec.model.enum.Instrument
 import seqexec.model.enum.Instrument._
 import mouse.boolean._
