+++ modules/seqexec/web/client/src/main/scala/seqexec/web/client/components/sequence/steps/OffsetDisplayCell.scala
@@ -5,7 +5,8 @@ package seqexec.web.client.components.sequence.steps
 
 import cats.Eq
 import seqexec.model.enum.Guiding
-import seqexec.model.Model.{ Offset, OffsetAxis, Step, TelescopeOffset }
+import seqexec.model.Step
+import seqexec.model.Model.{ Offset, OffsetAxis, TelescopeOffset }
 import seqexec.web.client.lenses.{telescopeOffsetPO, telescopeOffsetQO}
 import seqexec.web.client.lenses._
 import seqexec.web.client.components.SeqexecStyles
