+++ modules/seqexec/web/client/src/main/scala/seqexec/web/client/components/sequence/steps/StepsControlButtons.scala
@@ -11,7 +11,6 @@ import japgolly.scalajs.react.{Callback, CallbackTo, ScalaComponent, CatsReact}
 import gem.Observation
 import mouse.all._
 import seqexec.model._
-import seqexec.model.Model._
 import seqexec.model.enum._
 import seqexec.model.operations.ObservationOperations._
 import seqexec.model.operations._
