+++ modules/seqexec/web/client/src/main/scala/seqexec/web/client/model/ModelOps.scala
@@ -12,12 +12,8 @@ import seqexec.model.enum.{
   Instrument,
   Resource
 }
-import seqexec.model.{ StepState, SequenceState }
-import seqexec.model.Model.{
-  SequenceView,
-  StandardStep,
-  Step
-}
+import seqexec.model.{ StepState, SequenceState, Step, StandardStep }
+import seqexec.model.Model.SequenceView
 
 /**
   * Contains useful operations for the seqexec model
